Rename stash_api
